+++ src/api/profile.ts
@@ -1,5 +1,43 @@
 import type { UserProfile } from '../types/user';
 
+export const myProfile = async (token: string) => {
+  try {
+    const response = await fetch(
+      'https://waffle-instaclone.kro.kr/api/user/profile',
+      {
+        headers: {
+          Authorization: `Bearer ${token}`,
+          accept: 'application/json',
+        },
+      },
+    );
+
+    if (response.ok) {
+      const profileData = (await response.json()) as UserProfile;
+      return profileData;
+    }
+    throw new Error('Profile fetch failed');
+  } catch (err) {
+    console.error('Profile fetch error:', err);
+    return null;
+  }
+};
+
+export async function fetchUserProfile(username: string): Promise<UserProfile> {
+  try {
+    const response = await fetch(
+      `https://waffle-instaclone.kro.kr/api/user/${username}`,
+    );
+    if (!response.ok) {
+      throw new Error('Failed to fetch user data');
+    }
+    return (await response.json()) as UserProfile;
+  } catch (error) {
+    console.error('Error fetching user data:', error);
+    throw error;
+  }
+}
+
 type ProfileUpdateData = {
   username?: string;
   introduce?: string;
