+++ src/pages/LoginPage.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
 
-import { signin } from '../api/singin';
+import { signin } from '../api/auth';
 import type { UserProfile } from '../types/user';
 
 type LoginPageProps = {
